requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cytoolz==0.12.3
decorator==5.1.1
Deprecated==1.2.14
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Session==0.8.0
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
hexbytes==0.3.1
idna==3.7
importlib_resources==6.4.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
limits==3.13.0
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
ordered-set==4.1.0
packaging==24.1
parsimonious==0.10.0
protobuf==5.27.3
py==1.11.0
pycryptodome==3.20.0
Pygments==2.18.0
python-dotenv==1.0.1
python-logging-loki==0.3.1
pyunormalize==15.1.0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retry==0.9.2
rfc3339==6.2
rich==13.7.1
rlp==4.0.1
rpds-py==0.20.0
soupsieve==2.6
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
web3==6.20.2
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
